requirements.txt
discord.py
PyNaCl
python-dotenv
googleapi
yt_dlp
pygal
svglib
redditeasy
pillow